+++ src/app/content/highlights/utils/paginationUtils.ts
@@ -35,13 +35,13 @@ const totalOfCountsPerSource: (counts: CountsPerSource) => number = flow(
 export const getNextPageSources = (
   remainingCounts: CountsPerSource,
   tree: ArchiveTree,
-  nextPageSize: number
+  nextPageSize?: number
 ): string[] => {
   // remainingCounts is not ordered, so starting with this to make sure we load pages sequentially
   const pages = findTreePages(tree);
 
   const reduceUntilPageSize = reduceUntil(
-    (counts: CountsPerSource) => totalOfCountsPerSource(counts) >= nextPageSize
+    (counts: CountsPerSource) => nextPageSize ? totalOfCountsPerSource(counts) >= nextPageSize : false
   );
 
   const addPageCount = (counts: CountsPerSource, page: LinkedArchiveTreeSection) => {
